==> notebooks/modeling/Step1_Simple_File_Import.ipynb <==
# Import all packages

from __future__ import division
import pandas as pd
import numpy as np
from collections import Counter
import matplotlib.pyplot as plt
%matplotlib inline

# Import long form of data

datapath = '../DataProcessing/MergedData/NGS_GDS_abundance.csv'

df = pd.read_csv(datapath)
print(df.shape)
print(df.columns)

# Quick data exploration

# race data
race_features = ['Race_Asian', 'Race_Native_American', 'Race_Black',
                 'Race_Native_Hawaiian', 'Race_White', 'Race_Multiple', 
                 'Race_Unknown']
race_df = df[race_features] == 'Checked'
print(race_df.sum(),'\n')

# exposure data
exposure_features = ['Exposure_MSM', 'Exposure_IDU', 'Exposure_blood_transfusion',
                     'Exposure_heterosexual', 'Exposure_hemophilia', 'Exposure_perinatal',
                     'Exposure_MSM_and_IDU', 'Exposure_heterosexual_and_IDU', 
                     'Exposure_other', 'Exposure_unknown']
exposure_df = df[exposure_features] == 'Checked'
exposure_df.head()
print(exposure_df.sum())

# Select desired features

datapath = '../DataProcessing/MergedData/NGS_GDS_abundance.csv'
df = pd.read_csv(datapath)

C = ['Patient','Visit','DateOfVisit','Age','Gender','ART','VL','iVL','pVL',
     'CD4','iCD4','nCD4','CD8','iCD8','nCD8','TMHDS','VisitDate','Years_seropositive',
     'Race_Black','Race_Native_Hawaiian','Race_White','Race_Multiple',
     'Exposure_MSM', 'Exposure_IDU','Exposure_blood_transfusion','Exposure_heterosexual',
     'Exposure_heterosexual_and_IDU','Exposure_other',
     'GDS','Prot','AAPos','Coverage',
     'A','R','N','D','C','Q','E','G','H','I','L','K','M','F','P','S','T','W','Y','V']

df = df[C]
df.head()

# Define functions to reformat data

def slice_position(df, pos):
    m = df['AAPos'] == pos
    sliced_df = df[m]
    return sliced_df

def reformat_section(df):
    '''Must be a dataframe containing only 1 position'''
    pos = str(int(list(df.AAPos)[0]))
    AAs = ['A','R','N','D','C','Q','E','G','H','I',
           'L','K','M','F','P','S','T','W','Y','V']
    newnames_dict = {}
    for item in AAs:
        newnames_dict[item]= pos+item
    df_renamed = df.rename(columns = newnames_dict)
    df_renamed.drop(['AAPos','Coverage', 'Prot', 'DateOfVisit', 'VisitDate'], axis=1, inplace=True)
    return df_renamed

# Reformat data into wide form

tat_length = 101
on_columns = ['Patient', 'Visit', 'Age', 'Gender', 'ART', 'VL','iVL', 'pVL', 'CD4', 'iCD4', 
              'nCD4', 'CD8', 'iCD8', 'nCD8', 'Years_seropositive', 'TMHDS', 'GDS',
              'Race_Black','Race_Native_Hawaiian','Race_White','Race_Multiple',
              'Exposure_MSM', 'Exposure_IDU','Exposure_blood_transfusion','Exposure_heterosexual',
              'Exposure_heterosexual_and_IDU','Exposure_other']

start_df = slice_position(df, 1)
sklearn_df = reformat_section(start_df)
for i in range(2,tat_length+1):
    pos_df = slice_position(df, i)
    pos_df2 = reformat_section(pos_df)
    sklearn_df = pd.merge(sklearn_df, pos_df2,
               left_on=on_columns,
               right_on=on_columns,
               how='outer') 
    
print(sklearn_df.shape)
sklearn_df.head()

# Race and Exposure summary

race_features_2 = ['Race_Black','Race_Native_Hawaiian', 'Race_White', 'Race_Multiple']
rdf2 = sklearn_df[race_features_2] == 'Checked'
print(rdf2.sum(),'\n')

exposure_features_2 = ['Exposure_MSM', 'Exposure_IDU','Exposure_blood_transfusion',
                       'Exposure_heterosexual','Exposure_heterosexual_and_IDU','Exposure_other']
edf2 = sklearn_df[exposure_features_2] == 'Checked'
print(edf2.sum())

# Designate genetic-only and full data sets

all_features = list(sklearn_df.columns)

#nongenetic_features = ['Patient','Visit','Age','Gender','ART','VL','iVL','pVL','CD4',
#     'iCD4','nCD4','CD8','iCD8','nCD8','TMHDS','GDS', 'GDS_Threshold']

nongenetic_features = ['Patient', 'Visit', 'Age', 'Gender', 'ART', 'VL','iVL', 'pVL', 'CD4', 'iCD4', 
              'nCD4', 'CD8', 'iCD8', 'nCD8', 'Years_seropositive', 'TMHDS', 'GDS',
              'Race_Black','Race_Native_Hawaiian','Race_White','Race_Multiple',
              'Exposure_MSM', 'Exposure_IDU','Exposure_blood_transfusion','Exposure_heterosexual',
              'Exposure_heterosexual_and_IDU','Exposure_other']

genetic_feature_list = [x for x in all_features if x not in nongenetic_features]
len(genetic_feature_list)

genetic_df = sklearn_df[genetic_feature_list]
#genetic_df = genetic_df.fillna(genetic_df.mean())
print(genetic_df.shape)
genetic_df.head()

sklearn_df.to_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_df.csv',index=False)
genetic_df.to_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/genetic_df.csv',index=False)
