==> notebooks/modeling/Step2_Preprocessing.ipynb <==
# Import packages

import pandas as pd
import numpy as np

# Check dataframes

sklearn_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_df.csv')
genetic_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/genetic_df.csv')
print sklearn_df.shape
print genetic_df.shape

genetic_list = sklearn_df.columns[27:]
genetic_list

# Convert "checked" to 1/0

clin_cols = sklearn_df.columns[:27]
sklearn_df = sklearn_df.replace(to_replace='Unchecked',value=0)
sklearn_df = sklearn_df.replace(to_replace='Checked',value=1)
sklearn_df[clin_cols].head()

# Impute NaNs with mean

sklearn_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_df.csv')
genetic_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/genetic_df.csv')
sklearn_df = sklearn_df.replace(to_replace='Unchecked',value=0)
sklearn_df = sklearn_df.replace(to_replace='Checked',value=1)
print sklearn_df.shape
print genetic_df.shape

clin_cols = ['Age', 'VL', 'iVL', 'pVL', 'CD4', 'iCD4','nCD4', 'CD8', 'iCD8', 'nCD8', 'TMHDS',
             'Years_seropositive','Race_Black', 'Race_Native_Hawaiian', 'Race_White', 'Race_Multiple',
             'Exposure_MSM', 'Exposure_IDU', 'Exposure_blood_transfusion', 'Exposure_heterosexual',
             'Exposure_heterosexual_and_IDU','Exposure_other', 'GDS']

for col_name in clin_cols:
    sklearn_df[col_name].fillna(round(sklearn_df[col_name].mean()), inplace=True)
    
genetic_list = sklearn_df.columns[27:]
for c in genetic_list:
    sklearn_df[c].fillna(round(sklearn_df[c].mean(),4), inplace=True)
    
# perform log transformation on viral loads
def log10_transform(x):
    return np.log10(x)

VL_t = sklearn_df['VL'].apply(log10_transform).copy()
iVL_t = sklearn_df['iVL'].apply(log10_transform).copy()
pVL_t = sklearn_df['pVL'].apply(log10_transform).copy()
sklearn_df.loc[:,'log10_VL'] = VL_t
sklearn_df.loc[:,'log10_iVL'] = iVL_t
sklearn_df.loc[:,'log10_pVL'] = pVL_t

# drop the transgender sample
transgender_mask = sklearn_df.Gender != 'Transgender'
sklearn_df = sklearn_df[transgender_mask]

# save to CSV
#sklearn_df.to_csv('/Users/Greg/Documents/Drexel/ML/sklearn_preprocessed_df1.csv', index=False)
sklearn_df.to_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_preprocessed_df1.csv',
                  index=False)

# Impute NaNs with median

sklearn_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_df.csv')
genetic_df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/genetic_df.csv')
sklearn_df = sklearn_df.replace(to_replace='Unchecked',value=0)
sklearn_df = sklearn_df.replace(to_replace='Checked',value=1)
print sklearn_df.shape
print genetic_df.shape

clin_cols = ['Age', 'VL', 'iVL', 'pVL', 'CD4', 'iCD4','nCD4', 'CD8', 'iCD8', 'nCD8', 'TMHDS',
             'Years_seropositive','Race_Black', 'Race_Native_Hawaiian', 'Race_White', 'Race_Multiple',
             'Exposure_MSM', 'Exposure_IDU', 'Exposure_blood_transfusion', 'Exposure_heterosexual',
             'Exposure_heterosexual_and_IDU','Exposure_other', 'GDS']

for col_name in clin_cols:
    sklearn_df[col_name].fillna(round(sklearn_df[col_name].median()), inplace=True)
    
genetic_list = sklearn_df.columns[27:]
for c in genetic_list:
    sklearn_df[c].fillna(round(sklearn_df[c].median(),4), inplace=True)
    
# perform log transformation on viral loads
def log10_transform(x):
    return np.log10(x)

VL_t = sklearn_df['VL'].apply(log10_transform).copy()
iVL_t = sklearn_df['iVL'].apply(log10_transform).copy()
pVL_t = sklearn_df['pVL'].apply(log10_transform).copy()
sklearn_df.loc[:,'log10_VL'] = VL_t
sklearn_df.loc[:,'log10_iVL'] = iVL_t
sklearn_df.loc[:,'log10_pVL'] = pVL_t

# drop the transgender sample
transgender_mask = sklearn_df.Gender != 'Transgender'
sklearn_df = sklearn_df[transgender_mask]

# save to CSV
#sklearn_df.to_csv('/Users/Greg/Documents/Drexel/ML/sklearn_preprocessed_df1.csv', index=False)
sklearn_df.to_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_preprocessed_df2.csv',
                  index=False)

print sklearn_df.shape
print genetic_df.shape

for c in list(sklearn_df.columns):
    print c
