==> notebooks/modeling/Step4_Clinical_Model_Prep.ipynb <==
# Import packages

from __future__ import division
import pandas as pd
import numpy as np
import tat_neuro_functions as tnf
from sklearn.linear_model import LogisticRegression
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# Import data

df = pd.read_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/sklearn_preprocessed_df2.csv')

clin_cols = ['GDS','TMHDS','Age','Gender','log10_VL','log10_pVL','CD4','nCD4','CD8','nCD8',
             'Years_seropositive','Race_Black', 'Race_Native_Hawaiian', 'Race_White', 'Race_Multiple',
             'Exposure_MSM', 'Exposure_IDU', 'Exposure_blood_transfusion', 'Exposure_heterosexual',
             'Exposure_heterosexual_and_IDU','Exposure_other']

clin_df = tnf.select_dataframe_columns(df, clin_cols)
clin_df.head()

# Prepare for pipeline

target = 'GDS'
threshold = 0.5
features = ['Age','Gender','log10_VL','log10_pVL','CD4','nCD4','CD8','nCD8','TMHDS',
            'Years_seropositive','Race_Black', 'Race_Native_Hawaiian', 'Race_White', 'Race_Multiple',
            'Exposure_MSM', 'Exposure_IDU', 'Exposure_blood_transfusion', 'Exposure_heterosexual',
            'Exposure_heterosexual_and_IDU','Exposure_other']

X_df, y = tnf.run_preparation_pipeline2(clin_df, features, target, threshold)
print y

# Impaired and nonimpaired split
print 'Impaired:',sum(y==1)
print 'Nonimpaired:',sum(y==0)

X_df.head()

# Variance Filtering

from sklearn.feature_selection import VarianceThreshold

selector = VarianceThreshold(threshold=0.1)
T = selector.fit_transform(X_df)

#print X_df.shape
#print selector.variances_
#print T.shape

keep_col_list = []
for c in X_df.columns:
    v = np.var(list(X_df[c]))
    if v>0.1:
        keep_col_list.append(c)

print keep_col_list

X_df = pd.DataFrame(T)
X_df.columns = keep_col_list

# Impaired and nonimpaired split
print 'Impaired:',sum(y==1)
print 'Nonimpaired:',sum(y==0)

print X_df.shape
X_df.head()

# Normalize

X_df = tnf.scale_dataframe(X_df)

# Impaired and nonimpaired split
print 'Impaired:',sum(y==1)
print 'Nonimpaired:',sum(y==0)

print X_df.shape
X_df.head()

# Save data
X_df['GDS'] = y
clin_clean_df = X_df
print clin_clean_df.shape

clin_clean_df.to_csv('/Users/Greg/Projects/HIV_Neuro_NGS/DataModeling/dataframes/clin_clean_df.csv', index=False)
clin_clean_df.head()
